third_party/blink/web_tests/external/wpt/css/cssom/HTMLLinkElement-disabled-002.tentative-expected.txt

This is a testharness.js-based test.
FAIL HTMLLinkElement.disabled reflects the <link disabled> attribute, and behaves consistently, when the sheet is an alternate assert_equals: expected null but got Element node <link title="alt" rel="alternate stylesheet" href="data:t...
Harness: the test ran to completion.
